tools/teams/requirements.txt
click
pyyaml
PyGithub